requirements.txt
# Mnemosyne Core V3.0 - Python Dependencies (Infrastructure Layer)
# Python 3.12.7

# Database
aiosqlite==0.19.0

# HTTP Client for Ollama API
httpx>=0.26.0

# System & VRAM Guard
pynvml==11.5.0
psutil==5.9.8

# UI Automation
uiautomation==2.0.20

# OCR
pydantic
redis>=5.0.0
pytesseract==0.3.10
redis>=5.0.0
requests

# Image Processing
Pillow>=10.4.0

# Optional GPU OCR (install if VRAM available)
# easyocr>=1.7.1

# Vision & AI - Cognitive Layer
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
protobuf>=4.25.0
sentencepiece>=0.1.99

# LLM Integration (Ollama for text inference)
ollama>=0.3.0

# Windows specific
pywin32>=306

# Status Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0

# ============================================================
# Phase 8: Graph RAG - LlamaIndex + NetworkX
# ============================================================
llama-index-core>=0.10.0
llama-index-vector-stores-redis>=0.1.0
llama-index-embeddings-ollama>=0.1.0
networkx>=3.0
